refactor: hooks folder to shared
