--- src/app/hooks/index.ts
@@ -1,6 +1,0 @@
-export { useCourseByTitle } from './use-course-by-title';
-export { useDataByName } from './use-data-by-name';
-export { useNearestCourse } from './use-nearest-course';
-export { usePositionDropdown } from './use-position-dropdown';
-export { useTitle } from './use-title';
-export { useWindowSize } from './use-window-size';
