+++ src/features/about/about.tsx
@@ -1,9 +1,10 @@
 /* eslint-disable @stylistic/jsx-one-expression-per-line */
 import { contentMap } from './about.data';
 import { InfoGrid } from './components';
-import { useCourseByTitle } from '@/app/hooks';
+
 import { type Course } from '@/app/types';
 import { ArrowIcon } from '@/icons';
+import { useCourseByTitle } from '@/shared/hooks/use-course-by-title';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Title } from '@/shared/ui/title';
