+++ src/features/course-main/course-main.tsx
@@ -1,11 +1,12 @@
 import { useLoaderData } from 'react-router-dom';
 import { getCourseStatus } from './utils';
 import { DateLang } from '@/app/components';
-import { useTitle } from '@/app/hooks';
-import { selectCourse } from '@/app/hooks/use-course-by-title/utils/select-course.ts';
+
 import { Course } from '@/app/types';
 import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
+import { selectCourse } from '@/shared/hooks/use-course-by-title/utils/select-course';
+import { useTitle } from '@/shared/hooks/use-title';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { SectionLabel } from '@/shared/ui/section-label';
 import { Subtitle } from '@/shared/ui/subtitle';
