+++ src/features/school/alumni/alumni.tsx
@@ -1,6 +1,7 @@
 import { alumni } from './constants';
-import { useWindowSize } from '@/app/hooks';
+
 import Image from '@/features/image';
+import { useWindowSize } from '@/shared/hooks/use-window-size';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { Title } from '@/shared/ui/title';
