+++ src/features/training-program/training-program.test.tsx
@@ -3,9 +3,9 @@ import { type Mock, beforeEach, vi } from 'vitest';
 import { TrainingProgram } from './training-program';
 import { renderWithRouter } from '@/__tests__/utils';
 import { ROUTES } from '@/app/const';
-import { useCourseByTitle } from '@/app/hooks';
+import { useCourseByTitle } from '@/shared/hooks/use-course-by-title';
 
-vi.mock('@/app/hooks');
+vi.mock('@/shared/hooks/use-course-by-title');
 
 describe('TrainingProgram', () => {
   describe('with "angular" props', () => {
