+++ src/pages/courses.tsx
@@ -1,7 +1,8 @@
 import { FC } from 'react';
-import { useTitle } from '@/app/hooks';
+
 import { General, Main, Courses as RSCourses } from '@/features/courses';
 import { StudyPath } from '@/features/study-path';
+import { useTitle } from '@/shared/hooks/use-title';
 import { Breadcrumbs } from '@/shared/ui/breadcrumbs';
 
 export const Courses: FC = () => {
