+++ src/pages/home.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { Places } from '@/app/components';
-import { useTitle } from '@/app/hooks/use-title/use-title';
+
 import { config } from '@/config';
 import {
   About,
@@ -14,6 +14,7 @@ import {
   Speakers,
   Support,
 } from '@/features/home';
+import { useTitle } from '@/shared/hooks/use-title';
 import { Breadcrumbs } from '@/shared/ui/breadcrumbs';
 
 export const Home: FC = () => {
