+++ src/shared/hooks/use-course-by-title/use-course-by-title.test.tsx
