+++ src/shared/hooks/use-course-by-title/utils/is-course.ts
