+++ src/shared/hooks/use-nearest-course/types.ts
