+++ src/shared/hooks/use-nearest-course/utils/choose-nearest-course.ts
