+++ src/shared/hooks/use-nearest-course/utils/define-prev-next.ts
