+++ src/shared/hooks/use-nearest-course/utils/fined-nearest-course.ts
